OpenSSL EdDSA: reuse digest signature base with nullable hash (one-shot path); JDK: unify RAW SPKI/PKCS#8 unwrap via helper; wire OID constants
