+++ cryptography-providers/openssl3/api/src/commonMain/kotlin/algorithms/Openssl3EdDSA.kt
@@ -125,180 +125,15 @@ internal object Openssl3EdDSA : EdDSA {
     }
 }
 
-@OptIn(ExperimentalNativeApi::class, UnsafeNumber::class)
+
 private class EdDsaSignatureGenerator(
     private val privateKey: CPointer<EVP_PKEY>,
-) : SignatureGenerator {
-    private val cleaner = privateKey.upRef().cleaner()
-
-    override fun createSignFunction(): SignFunction = EdDsaSignFunction(privateKey)
-
-    override fun generateSignatureBlocking(data: ByteArray): ByteArray = memScoped {
-        val ctx = checkError(EVP_MD_CTX_new())
-        try {
-            checkError(
-                EVP_DigestSignInit_ex(
-                    ctx = ctx,
-                    pctx = null,
-                    mdname = null, // must be null for EdDSA one-shot
-                    libctx = null,
-                    props = null,
-                    pkey = privateKey,
-                    params = null
-                )
-            )
-
-            data.usePinned {
-                val siglen = alloc<size_tVar>()
-                checkError(EVP_DigestSign(ctx, null, siglen.ptr, it.safeAddressOfU(0), data.size.convert()))
-                val out = ByteArray(siglen.value.convert())
-                out.usePinned { outPin ->
-                    checkError(EVP_DigestSign(ctx, outPin.safeAddressOfU(0), siglen.ptr, it.safeAddressOfU(0), data.size.convert()))
-                }
-                out.ensureSizeExactly(siglen.value.convert())
-            }
-        } finally {
-            EVP_MD_CTX_free(ctx)
-        }
-    }
+) : Openssl3DigestSignatureGenerator(privateKey, hashAlgorithm = null) {
+    override fun MemScope.createParams(): CValuesRef<OSSL_PARAM>? = null
 }
 
-@OptIn(ExperimentalNativeApi::class, UnsafeNumber::class)
 private class EdDsaSignatureVerifier(
     private val publicKey: CPointer<EVP_PKEY>,
-) : SignatureVerifier {
-    private val cleaner = publicKey.upRef().cleaner()
-
-    override fun createVerifyFunction(): VerifyFunction = EdDsaVerifyFunction(publicKey)
-}
-
-@OptIn(UnsafeNumber::class)
-private class EdDsaSignFunction(
-    private val privateKey: CPointer<EVP_PKEY>,
-) : SignFunction {
-    private var isClosed = false
-    private var accumulator = EmptyByteArray
-
-    private fun ensureOpen() = check(!isClosed) { "Already closed" }
-
-    override fun update(source: ByteArray, startIndex: Int, endIndex: Int) {
-        ensureOpen()
-        checkBounds(source.size, startIndex, endIndex)
-        accumulator += source.copyOfRange(startIndex, endIndex)
-    }
-
-    override fun signIntoByteArray(destination: ByteArray, destinationOffset: Int): Int {
-        val sig = signToByteArray()
-        sig.copyInto(destination, destinationOffset)
-        return sig.size
-    }
-
-    override fun signToByteArray(): ByteArray {
-        ensureOpen()
-        return memScoped {
-            val ctx = checkError(EVP_MD_CTX_new())
-            try {
-                checkError(
-                    EVP_DigestSignInit_ex(
-                        ctx = ctx,
-                        pctx = null,
-                        mdname = null,
-                        libctx = null,
-                        props = null,
-                        pkey = privateKey,
-                        params = null
-                    )
-                )
-                accumulator.usePinned { pin ->
-                    val siglen = alloc<size_tVar>()
-                    checkError(EVP_DigestSign(ctx, null, siglen.ptr, pin.safeAddressOfU(0), accumulator.size.convert()))
-                    val out = ByteArray(siglen.value.convert())
-                    out.usePinned { outPin ->
-                        checkError(EVP_DigestSign(ctx, outPin.safeAddressOfU(0), siglen.ptr, pin.safeAddressOfU(0), accumulator.size.convert()))
-                    }
-                    out.ensureSizeExactly(siglen.value.convert())
-                }
-            } finally {
-                EVP_MD_CTX_free(ctx)
-                reset()
-            }
-        }
-    }
-
-    override fun reset() {
-        ensureOpen()
-        accumulator = EmptyByteArray
-    }
-
-    override fun close() {
-        isClosed = true
-        accumulator = EmptyByteArray
-    }
-}
-
-@OptIn(UnsafeNumber::class)
-private class EdDsaVerifyFunction(
-    private val publicKey: CPointer<EVP_PKEY>,
-) : VerifyFunction {
-    private var isClosed = false
-    private var accumulator = EmptyByteArray
-
-    private fun ensureOpen() = check(!isClosed) { "Already closed" }
-
-    override fun update(source: ByteArray, startIndex: Int, endIndex: Int) {
-        ensureOpen()
-        checkBounds(source.size, startIndex, endIndex)
-        accumulator += source.copyOfRange(startIndex, endIndex)
-    }
-
-    override fun tryVerify(signature: ByteArray, startIndex: Int, endIndex: Int): Boolean {
-        ensureOpen()
-        checkBounds(signature.size, startIndex, endIndex)
-        return memScoped {
-            val ctx = checkError(EVP_MD_CTX_new())
-            try {
-                checkError(
-                    EVP_DigestVerifyInit_ex(
-                        ctx = ctx,
-                        pctx = null,
-                        mdname = null,
-                        libctx = null,
-                        props = null,
-                        pkey = publicKey,
-                        params = null
-                    )
-                )
-                signature.usePinned { sigPin ->
-                    accumulator.usePinned { dataPin ->
-                        val res = EVP_DigestVerify(
-                            ctx,
-                            sigPin.safeAddressOfU(startIndex),
-                            (endIndex - startIndex).convert(),
-                            dataPin.safeAddressOfU(0),
-                            accumulator.size.convert()
-                        )
-                        if (res != 0) checkError(res)
-                        res == 1
-                    }
-                }
-            } finally {
-                EVP_MD_CTX_free(ctx)
-                reset()
-            }
-        }
-    }
-
-    override fun verify(signature: ByteArray, startIndex: Int, endIndex: Int) {
-        check(tryVerify(signature, startIndex, endIndex)) { "Invalid signature" }
-    }
-
-    override fun reset() {
-        ensureOpen()
-        accumulator = EmptyByteArray
-    }
-
-    override fun close() {
-        isClosed = true
-        accumulator = EmptyByteArray
-    }
+) : Openssl3DigestSignatureVerifier(publicKey, hashAlgorithm = null) {
+    override fun MemScope.createParams(): CValuesRef<OSSL_PARAM>? = null
 }
