+++ cryptography-providers/openssl3/api/src/commonMain/kotlin/operations/Openssl3DigestSignatureGenerator.kt
@@ -14,19 +14,21 @@ import kotlin.experimental.*
 
 internal abstract class Openssl3DigestSignatureGenerator(
     private val privateKey: CPointer<EVP_PKEY>,
-    private val hashAlgorithm: String,
+    // when null, performs one-shot signing (e.g., EdDSA)
+    private val hashAlgorithm: String?,
 ) : SignatureGenerator {
     @OptIn(ExperimentalNativeApi::class)
     private val cleaner = privateKey.upRef().cleaner()
 
     protected abstract fun MemScope.createParams(): CValuesRef<OSSL_PARAM>?
 
-    override fun createSignFunction(): SignFunction {
-        return Openssl3DigestSignFunction(Resource(checkError(EVP_MD_CTX_new()), ::EVP_MD_CTX_free))
+    override fun createSignFunction(): SignFunction = when (hashAlgorithm) {
+        null -> OneShotSignFunction()
+        else -> StreamingSignFunction(Resource(checkError(EVP_MD_CTX_new()), ::EVP_MD_CTX_free))
     }
 
     // inner class to have a reference to class with cleaner
-    private inner class Openssl3DigestSignFunction(
+    private inner class StreamingSignFunction(
         private val context: Resource<CPointer<EVP_MD_CTX>>,
     ) : SignFunction, SafeCloseable(SafeCloseAction(context, AutoCloseable::close)) {
         init {
@@ -66,13 +68,77 @@ internal abstract class Openssl3DigestSignatureGenerator(
                 EVP_DigestSignInit_ex(
                     ctx = context,
                     pctx = null,
-                    mdname = hashAlgorithm,
+                    mdname = hashAlgorithm!!,
                     libctx = null,
                     props = null,
                     pkey = privateKey,
                     params = createParams()
                 )
             )
+        }
+    }
+
+    private inner class OneShotSignFunction : SignFunction {
+        private var isClosed = false
+        private var accumulator = EmptyByteArray
+
+        private fun ensureOpen() = check(!isClosed) { "Already closed" }
+
+        override fun update(source: ByteArray, startIndex: Int, endIndex: Int) {
+            ensureOpen()
+            checkBounds(source.size, startIndex, endIndex)
+            // accumulate until final
+            accumulator += source.copyOfRange(startIndex, endIndex)
+        }
+
+        override fun signIntoByteArray(destination: ByteArray, destinationOffset: Int): Int {
+            val sig = signToByteArray()
+            checkBounds(destination.size, destinationOffset, destinationOffset + sig.size)
+            sig.copyInto(destination, destinationOffset)
+            return sig.size
+        }
+
+        @OptIn(UnsafeNumber::class)
+        override fun signToByteArray(): ByteArray = memScoped {
+            ensureOpen()
+            val ctx = checkError(EVP_MD_CTX_new())
+            try {
+                checkError(
+                    EVP_DigestSignInit_ex(
+                        ctx = ctx,
+                        pctx = null,
+                        mdname = null, // one-shot mode
+                        libctx = null,
+                        props = null,
+                        pkey = privateKey,
+                        params = createParams()
+                    )
+                )
+
+                val siglen = alloc<size_tVar>()
+                accumulator.usePinned { pin ->
+                    checkError(EVP_DigestSign(ctx, null, siglen.ptr, pin.safeAddressOfU(0), accumulator.size.convert()))
+                    val out = ByteArray(siglen.value.convert())
+                    out.usePinned { outPin ->
+                        checkError(EVP_DigestSign(ctx, outPin.safeAddressOfU(0), siglen.ptr, pin.safeAddressOfU(0), accumulator.size.convert()))
+                    }
+                    out.ensureSizeExactly(siglen.value.convert())
+                    out
+                }
+            } finally {
+                EVP_MD_CTX_free(ctx)
+                isClosed = true
+            }
+        }
+
+        override fun reset() {
+            isClosed = false
+            accumulator = EmptyByteArray
+        }
+
+        override fun close() {
+            isClosed = true
+            accumulator = EmptyByteArray
         }
     }
 }
